+++ tests/tconvert.nim
@@ -15,7 +15,7 @@ proc create_dset(h5f: var H5FileObj): H5DataSet =
   result = h5f.create_dataset(DsetName, (2, 5), int)
   result[result.all] = d_ar
 
-when isMainModule:
+proc main =
   # open file, create dataset
   var
     h5f = H5open(File, "rw")
@@ -49,8 +49,9 @@ when isMainModule:
   assertType(float32)
   assertType(float64)
 
-  var err = h5f.close()
-  doAssert(err >= 0)
+  doAssert h5f.close() >= 0
 
+when isMainModule:
+  main()
   # clean up after ourselves
   removeFile(File)
